refactor: extract TimelineRow into a standalone component
